rename file header, footer e home
